Fix flaky distributed CI by moving benchmark_utils
* Move nvfuser/testing/benchmark_utils.py to tests/python/multidevice
* Direct bindings tests can avoid importing nvfuser
